+++ src/main/java/com/hypherionmc/modpublisher/Constants.java
@@ -16,6 +16,7 @@ public class Constants {
     public static final String CURSE_TASK = "publishCurseforge";
     public static final String GITHUB_TASK = "publishGitHub";
     public static final String MODRINTH_TASK = "publishModrinth";
+    public static final String NIGHTBLOOM_TASK = "publishNightbloom";
     public static final String TASK_GROUP = "publishing";
     public static final String EXTENSION_NAME = "publisher";